key-extract.py: gracefully handle missing IID for traffic secret
When the kprobe trace is truncated (dropped events?), the IID from
consume_initiation can be missing. Handle this case gracefully,
hopefully the initiator can print the secret anyway (in a configuration
using network namespaces).

It is probably better to spew less things in kprobes and find more
reliable ways to link keys.