README: rewrite key extraction sections
Remove uninteresting information (like the exact material that gets
extracted and how they are used) in favor of a user-centered approach
for key-probe.sh and key-extract.py.

The README is a bit long, so shrink the extract-keys section. The
tarball information is no longer neded since 0.0.20170214-1-g1c0ee32.
But with that change, the fixup for Arch Linux also does not work
anymore, so just remove it and refer to the WG sources for information.

The tarball information is no longer needed for newer version, so drop
that.